# spec/contracts/modeling/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Modeling API
  version: 1.0.0
  description: |
    Синхронный REST API для BC Modeling (docs/backend.md §5).

    Каждая операция ссылается на use case в `x-use-case` и на операцию таблицы
    доступа в `x-operation` (spec/nfr/modeling.yaml#security.authorization): роль
    проверяется на границе use case'а, а не матчером URL. Наполнено на этапе 2
    (2026-10-08) по одному endpoint'у на экспонированный сценарий; до этого
    `paths` был пуст намеренно — endpoint без `UC-*` нарушал бы трассируемость.

    Ошибки — `application/problem+json` (§8.1) с полем `code`: код инварианта
    (`INV-MDL-NNN`) или бизнес-ошибки (`RELATION_NOT_PERMITTED`, `MDL_*`).
    Класс отказа домена определяет статус: не найдено или скрыто списком доступа —
    `404`, конфликт состояния — `409`, нарушение правила — `422`, роль — `403`,
    очищенный снимок — `410`.

    Остаток XML (`raw_xml`, ADR-0017) наружу не отдаётся: это внутренний формат
    interchange.
servers:
  - url: /api/v1
    description: Относительный путь; приложение и фронтенд отдаются одним артефактом.
security:
  - bearerAuth: []
...
paths:
...
  /models:
    get:
      tags: [models]
      operationId: listModels
      summary: Модели пространства, видимые автору (FR-04, INV-MDL-011)
      x-use-case: UC-MDL-001
      x-operation: OpenModel
      parameters:
        - { in: query, name: workspaceId, required: false, schema: { type: string, format: uuid },
            description: "Без него — первое пространство." }
      responses:
        '200':
          description: Удалённые видит только ADMIN.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ModelSummary' } }
    post:
      tags: [models]
      operationId: createModel
      summary: Пустая модель с девятью корневыми папками и версией 1
      x-use-case: UC-MDL-001
      x-operation: CreateModel
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, maxLength: 500 }
                workspaceId: { type: string, format: uuid }
      responses:
        '201':
          description: Модель создана; повтор ключа с тем же телом — та же модель (INV-MDL-003).
          content:
            application/json: { schema: { $ref: '#/components/schemas/ModelSummary' } }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '409': { $ref: '#/components/responses/Conflict409' }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
  /models/{id}:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [models]
      operationId: openModel
      summary: Дерево модели — папки, элементы, связи, список представлений
      x-use-case: UC-MDL-001
      x-operation: OpenModel
      responses:
        '200':
          description: Списки — в порядке sort_order.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ModelTree' } }
        '404': { $ref: '#/components/responses/NotFound404' }
    patch:
      tags: [models]
      operationId: editModel
      summary: Переименование и документация
      x-use-case: UC-MDL-001
      x-operation: RenameModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:          { type: string, maxLength: 500 }
                documentation: { type: string, description: "Пустая строка снимает документацию." }
      responses:
        '200':
          description: Изменено.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ModelSummary' } }
        '409': { $ref: '#/components/responses/Conflict409' }
    delete:
      tags: [models]
      operationId: deleteModel
      summary: Soft delete (INV-MDL-002); требует своей блокировки
      x-use-case: UC-MDL-001
      x-operation: DeleteModel
      responses:
        '204': { description: Модель в DELETED, блокировка снята. }
        '409': { $ref: '#/components/responses/Conflict409' }
  /models/{id}/restore:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [models]
      operationId: restoreModel
      summary: DELETED → ACTIVE, только ADMIN
      x-use-case: UC-MDL-001
      x-operation: RestoreModel
      responses:
        '200':
          description: Восстановлена.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ModelSummary' } }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
  /models/{id}/purge:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [models]
      operationId: purgeModel
      summary: DELETED → PURGED, только ADMIN; ACTIVE → PURGED запрещён (INV-MDL-002)
      x-use-case: UC-MDL-001
      x-operation: PurgeModel
      responses:
        '204': { description: Модель уничтожена вместе с версиями. }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
  /models/{id}/lock:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [models]
      operationId: currentLock
      summary: Действующая блокировка — чтобы остальные видели модель read-only
      x-use-case: UC-MDL-005
      x-operation: OpenModel
      responses:
        '200':
          description: Модель заблокирована.
          content:
            application/json: { schema: { $ref: '#/components/schemas/LockInfo' } }
        '204': { description: Блокировки нет. }
    post:
      tags: [models]
      operationId: acquireLock
      summary: Захват или продление своей блокировки (FR-05)
      x-use-case: UC-MDL-005
      x-operation: AcquireLock
      responses:
        '200':
          description: Блокировка у автора до expiresAt.
          content:
            application/json: { schema: { $ref: '#/components/schemas/LockInfo' } }
        '409': { $ref: '#/components/responses/Conflict409' }
    delete:
      tags: [models]
      operationId: releaseLock
      summary: Освобождение; чужую — только ADMIN
      x-use-case: UC-MDL-005
      x-operation: ReleaseLock
      responses:
        '204': { description: Снята или её не было. }
        '403': { $ref: '#/components/responses/Forbidden403' }
...
  /models/{id}/validate:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [models]
      operationId: validateModel
      summary: Отчёт валидации метамодели без ИИ — нарушения из импорта видны здесь (FR-10)
      x-use-case: UC-MDL-003
      x-operation: ValidateModel
      responses:
        '200':
          description: Находки, сначала ERROR.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Finding' } }
  /models/{id}/acl:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [models]
      operationId: getModelAccess
      summary: Список доступа (FR-29)
      x-use-case: UC-MDL-007
      x-operation: OpenModel
      responses:
        '200':
          description: Пустой — модель открыта по ролям.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AclEntry' } }
    put:
      tags: [models]
      operationId: replaceModelAccess
      summary: Заменить список целиком — автор модели или ADMIN (INV-MDL-011)
      x-use-case: UC-MDL-007
      x-operation: ManageModelAccess
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: array, items: { $ref: '#/components/schemas/AclEntry' } }
      responses:
        '200':
          description: Новый список.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AclEntry' } }
        '403': { $ref: '#/components/responses/Forbidden403' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT, выданный Keycloak. Роли VIEWER / ARCHITECT / ADMIN — в claim'ах токена (FR-28), группы — в `groups`.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, maxLength: 64 }
      description: |
        Ключ дедупликации повторов. Тот же ключ с тем же телом возвращает результат
        первого выполнения; тот же ключ с другим телом — `409` (INV-MDL-003).
        Ключ принадлежит пользователю: чужой ключ с тем же значением не конфликтует.
    ModelId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Problem:
      type: object
      description: |
        RFC 7807 Problem Details + поле `code` (docs/backend.md §8.1). Дополнительные поля
        по ситуации: `lockOwner` и `expiresAt` у чужой блокировки, `relationships` у удаления
        связанного элемента, `permitted` у недопустимой связи.
      required: [type, title, status, code]
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }
        code:
          type: string
          description: Код инварианта (`INV-MDL-NNN`, domain/modeling/invariants.md) или иной бизнес-ошибки.
          examples: ["INV-MDL-004", "INV-MDL-006", "RELATION_NOT_PERMITTED"]
        lockOwner:     { type: string }
        relationships: { type: array, items: { type: string, format: uuid } }
        permitted:     { type: array, items: { type: string } }
...
    ModelSummary:
      type: object
      properties:
        id:            { type: string, format: uuid }
        workspaceId:   { type: string, format: uuid }
        archiId:       { type: string }
        name:          { type: string }
        documentation: { type: string }
        status:        { type: string, enum: [ACTIVE, DELETED] }
        archiVersion:  { type: string }
        createdBy:     { type: string }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
    Property:
      type: object
      properties:
        key:   { type: string, maxLength: 200 }
        value: { type: string }
    Folder:
      type: object
      properties:
        id:         { type: string, format: uuid }
        parentId:   { type: string, format: uuid }
        archiId:    { type: string }
        name:       { type: string }
        folderType: { type: string, enum: [STRATEGY, BUSINESS, APPLICATION, TECHNOLOGY, MOTIVATION, IMPLEMENTATION_MIGRATION, OTHER, RELATIONS, DIAGRAMS] }
        sortOrder:  { type: integer, format: int64 }
    Element:
      type: object
      properties:
        id:            { type: string, format: uuid }
        folderId:      { type: string, format: uuid }
        archiId:       { type: string }
        archiType:     { type: string }
        layer:         { type: string }
        name:          { type: string }
        documentation: { type: string }
        properties:    { type: array, items: { $ref: '#/components/schemas/Property' } }
        supported:     { type: boolean, description: "false — opaque (FR-03): хранится, но не редактируется." }
        sortOrder:     { type: integer, format: int64 }
...
    Relationship:
      type: object
      properties:
        id:            { type: string, format: uuid }
        folderId:      { type: string, format: uuid }
        archiId:       { type: string }
        archiType:     { type: string }
        sourceId:      { type: string, format: uuid }
        targetId:      { type: string, format: uuid }
        name:          { type: string }
        documentation: { type: string }
        accessType:    { type: string, enum: [WRITE, READ, ACCESS, READ_WRITE] }
        directed:      { type: boolean }
        properties:    { type: array, items: { $ref: '#/components/schemas/Property' } }
        supported:     { type: boolean }
        sortOrder:     { type: integer, format: int64 }
        edgeId:        { type: string, format: uuid, description: "Ребро, если его просили создать." }
    ViewSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        folderId:  { type: string, format: uuid }
        archiId:   { type: string }
        archiType: { type: string }
        name:      { type: string }
        sortOrder: { type: integer, format: int64 }
    ModelTree:
      type: object
      properties:
        model:         { $ref: '#/components/schemas/ModelSummary' }
        folders:       { type: array, items: { $ref: '#/components/schemas/Folder' } }
        elements:      { type: array, items: { $ref: '#/components/schemas/Element' } }
        relationships: { type: array, items: { $ref: '#/components/schemas/Relationship' } }
        views:         { type: array, items: { $ref: '#/components/schemas/ViewSummary' } }
...
    LockInfo:
      type: object
      properties:
        modelId:    { type: string, format: uuid }
        owner:      { type: string }
        acquiredAt: { type: string, format: date-time }
        expiresAt:  { type: string, format: date-time }
    Finding:
      type: object
      description: Общий формат находки — валидатор и ИИ (§5.1).
      properties:
        severity:   { type: string, enum: [ERROR, WARNING, INFO] }
        code:       { type: string }
        message:    { type: string }
        targetKind: { type: string }
        targetId:   { type: string, description: "archi_id объекта." }
        suggestion: { type: string }
    AclEntry:
      type: object
      required: [principalType, principal, access]
      properties:
        principalType: { type: string, enum: [USER, GROUP] }
        principal:     { type: string, maxLength: 200 }
        access:        { type: string, enum: [READ, WRITE] }
...
  responses:
    NotFound404:
      description: Объекта нет — или он скрыт списком доступа модели (INV-MDL-011).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
    Forbidden403:
      description: Роль не даёт права на операцию (FR-28) либо список доступа даёт только чтение.
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
    Conflict409:
      description: |
        Блокировки нет или она чужая, конкурентное сохранение (INV-MDL-006), повтор
        Idempotency-Key с другим телом (INV-MDL-003), удаление связанного элемента
        (INV-MDL-004), команда к удалённой модели (INV-MDL-002).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
    UnprocessableEntity422:
      description: |
        Нарушение бизнес-правила: недопустимая связь (INV-MDL-007), тип не той фазы,
        папка чужой модели или не того корня (INV-MDL-009), запрещённый переход (INV-MDL-002).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
...